requirements.txt
Pillow
opencv-python
numpy
tqdm
# for cuda
cupy